+++ R/Gibbs_IMFA.R
@@ -38,7 +38,7 @@
     if(sw["pi.sw"])  {
       pi.store       <- matrix(0L, nrow=trunc.G, ncol=n.store)
     }
-    z.store          <- matrix(0L, nrow=N, ncol=n.store)
+    z.store          <- matrix(0L, nrow=n.store, ncol=N)
     ll.store         <- rep(0L, n.store)
     acc1             <- acc2 <- FALSE
     err.z            <- zerr <- FALSE
@@ -132,7 +132,7 @@
       if(sw["l.sw"])    load.store[,,,1] <- lmat
       if(sw["psi.sw"])  psi.store[,,1]   <- 1/psi.inv
       if(sw["pi.sw"])   pi.store[,1]     <- pi.prop
-      z.store[,1]            <- z
+      z.store[1,]            <- z
       sigma                  <- if(uni) lapply(Gs, function(g) as.matrix(1/psi.inv[,g] + if(Q0) tcrossprod(as.matrix(lmat[,,g])) else 0)) else lapply(Gs, function(g) tcrossprod(lmat[,,g]) + diag(1/psi.inv[,g]))
       log.probs              <- if(uni) vapply(Gs, function(g) dnorm(data, mu[,g], sqrt(sigma[[g]]), log=TRUE) + log(pi.prop[g]), numeric(N)) else vapply(Gs, function(g) { sigma <- if(Q0) sigma[[g]] else sqrt(sigma[[g]]); dmvn(data, mu[,g], is.posi_def(sigma, make=TRUE)$X.new, log=TRUE, isChol=!Q0) + log(pi.prop[g]) }, numeric(N))
       ll.store[1]            <- sum(rowLogSumExps(log.probs))
@@ -350,7 +350,7 @@
         if(sw["pi.sw"])               pi.store[Gs,new.it]     <- pi.prop
         if(learn.alpha)               alpha.store[new.it]     <- pi.alpha
         if(learn.d)                   d.store[new.it]         <- discount
-                                      z.store[,new.it]        <- as.integer(z)
+                                      z.store[new.it,]        <- as.integer(z)
                                       ll.store[new.it]        <- if(G > 1) sum(rowLogSumExps(log.probs)) else sum(dmvn(X=data, mu=mu[,nn.ind], sigma=tcrossprod(as.matrix(lmat[,,nn.ind])) + diag(psi.store[,nn.ind,new.it]), log=TRUE))
                                       G.store[new.it]         <- as.integer(G.non)
                                       act.store[new.it]       <- as.integer(G)