+++ R/Gibbs_IMIFA.R
@@ -37,7 +37,7 @@
     if(sw["pi.sw"])  {
       pi.store       <- matrix(0L, nrow=trunc.G, ncol=n.store)
     }
-    z.store          <- matrix(0L, nrow=N, ncol=n.store)
+    z.store          <- matrix(0L, nrow=n.store, ncol=N)
     ll.store         <- rep(0L, n.store)
     Q.star           <- Q
     Qs               <- rep(Q,  trunc.G)
@@ -140,7 +140,7 @@
       if(sw["l.sw"])    load.store[,,,1] <- array(unlist(lmat, use.names=FALSE), dim=c(P, Q, trunc.G))
       if(sw["psi.sw"])  psi.store[,,1]   <- 1/psi.inv
       if(sw["pi.sw"])   pi.store[,1]     <- pi.prop
-      z.store[,1]             <- z
+      z.store[1,]             <- z
       Q0                      <- Qs > 0
       sigma                   <- if(uni) lapply(Gs, function(g) as.matrix(1/psi.inv[,g] + if(Q0[g]) tcrossprod(lmat[[g]]) else 0)) else lapply(Gs, function(g) tcrossprod(lmat[[g]]) + diag(1/psi.inv[,g]))
       log.probs               <- if(uni) vapply(Gs, function(g) dnorm(data, mu[,g], sqrt(sigma[[g]]), log=TRUE) + log(pi.prop[g]), numeric(N)) else vapply(Gs, function(g, Q=Q0[g]) { sigma <- if(Q) sigma[[g]] else sqrt(sigma[[g]]); dmvn(data, mu[,g], is.posi_def(sigma, make=TRUE)$X.new, log=TRUE, isChol=!Q) + log(pi.prop[g]) }, numeric(N))
@@ -485,7 +485,7 @@
         if(sw["pi.sw"])  pi.store[Gs,new.it]           <- pi.prop
         if(learn.alpha)  alpha.store[new.it]           <- pi.alpha
         if(learn.d)      d.store[new.it]               <- discount
-                         z.store[,new.it]              <- as.integer(z)
+                         z.store[new.it,]              <- as.integer(z)
                          ll.store[new.it]              <- if(G > 1) sum(rowLogSumExps(log.probs)) else sum(dmvn(X=data, mu=mu[,nn.ind], sigma=tcrossprod(lmat[[nn.ind]]) + diag(psi.store[,nn.ind,new.it]), log=TRUE))
                          Q.store[,new.it]              <- as.integer(Qs)
                          G.store[new.it]               <- as.integer(G.non)